+++ app/Http/Controllers/Auth/RegisterController.php
@@ -74,6 +74,6 @@ class RegisterController extends Controller
 
     protected function redirectTo()
     {
-        return route('tops.question');
+        return route('questions.index');
     }
 }
